requirements.txt
flask>=3.0.0
google-generativeai>=0.7.0
chromadb>=0.5.0
python-dotenv>=1.0.0
gunicorn>=22.0.0
flask-cors>=4.0.0
flask-limiter>=3.5.0
